Attempt workaround for MSVC C4245 signed/unsigned mismatch in template 'argument'
https://github.com/upx/upx/issues/435
	modified:   pefile.cpp

+++ src/pefile.cpp
@@ -3110,7 +3110,9 @@ void PeFile32::pack0(OutputFile *fo, unsigned subsystem_mask,
                      upx_uint64_t default_imagebase,
                      bool last_section_rsrc_only)
 {
-    super::pack0<LE32>(fo, ih, oh, subsystem_mask,
+    // Redundant cast of subsystem_mask tries to workaround
+    // MSVC complaint C4245 signed/unsigned mismatch
+    super::pack0<LE32>(fo, ih, oh, (unsigned)subsystem_mask,
                        default_imagebase, last_section_rsrc_only);
 }
